+++ engine/lite/embeddedDeploymentManager/src/test/scala/pl/touk/nussknacker/streaming/embedded/RequestResponseEmbeddedDeploymentManagerTest.scala
@@ -7,6 +7,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.ProcessVersion
+import pl.touk.nussknacker.engine.api.deployment.cache.ScenarioStateCachingConfig
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.deployment.{
   DataFreshnessPolicy,
@@ -53,7 +54,8 @@ class RequestResponseEmbeddedDeploymentManagerTest extends AnyFunSuite with Matc
         .empty()
         .withValue("mode", fromAnyRef("request-response"))
         .withValue("http.port", fromAnyRef(port))
-        .withValue("http.interface", fromAnyRef("localhost"))
+        .withValue("http.interface", fromAnyRef("localhost")),
+      ScenarioStateCachingConfig.Default.cacheTTL
     )
     FixtureParam(manager, modelData, port)
   }